modify input data selector
